tests/requirements.txt
boto3==1.26.137
botocore==1.29.165
certifi==2025.4.26
charset-normalizer==3.4.2
eis-insurance==1.65.0
idna==3.10
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.2
six==1.17.0
urllib3==1.26.20
msal==1.25.0
msgraph-core==0.2.2
msgraph-sdk==1.0.0
reportlab==4.0.7